+++ src/interpreter/runtimeModel.ts
@@ -1,7 +1,8 @@
+import { WOLLOK_BASE_PACKAGE, WOLLOK_EXTRA_STACK_TRACE_HEADER } from '../constants'
 import { v4 as uuid } from 'uuid'
 import { getPotentiallyUninitializedLazy } from '../decorators'
 import { get, is, last, List, match, raise, when } from '../extensions'
-import { Assignment, Body, Catch, Describe, Environment, Entity, Expression, Id, If, Literal, LiteralValue, Method, Module, Name, New, Node, Package, Program, Reference, Return, Self, Send, Singleton, Super, Test, Throw, Try, Variable } from '../model'
+import { Assignment, Body, Catch, Describe, Environment, Entity, Expression, Id, If, Literal, LiteralValue, Method, Module, Name, New, Node, Package, Program, Reference, Return, Self, Send, Singleton, Super, Test, Throw, Try, Variable, Class } from '../model'
 import { Interpreter } from './interpreter'
 
 const { isArray } = Array
@@ -50,7 +51,7 @@ export class WollokException extends Error {
   get message(): string {
     const error: RuntimeObject = this.instance
     error.assertIsException()
-    return `${error.innerValue ? error.innerValue.message : error.get('message')?.innerString ?? ''}\n${this.wollokStack}\n     Derived from TypeScript stack`
+    return `${error.innerValue ? error.innerValue.message : error.get('message')?.innerString ?? ''}\n${this.wollokStack}\n     ${WOLLOK_EXTRA_STACK_TRACE_HEADER}`
   }
 
   // TODO: Do we need to take this into consideration for Evaluation.copy()? This might be inside Exception objects
@@ -151,6 +152,11 @@ export class Frame extends Context {
 
   override toString(): string {
     return `${this.description}(${this.sourceInfo})`
+  }
+
+  isCustom(): boolean {
+    const module = this.node.ancestors.find(ancestor => ancestor.is(Module)) as Module
+    return !module?.fullyQualifiedName?.startsWith(WOLLOK_BASE_PACKAGE) && !this.node.is(Environment)
   }
 }
 
@@ -481,6 +487,12 @@ export class Evaluation {
     yield node
 
     const target = node.instantiated.target ?? raise(new Error(`Could not resolve reference to instantiated module ${node.instantiated.name}`))
+
+    const name = node.instantiated.name
+    
+    if (!target.is(Class)) raise(new Error(`${name} is not a class, you cannot generate instances of a ${target?.kind}`))
+
+    if (target.isAbstract) raise(new Error(`${name} is an abstract class, you cannot generate instances`))
 
     return yield* this.instantiate(target, args)
   }