+++ test/validator.test.ts
@@ -4,9 +4,9 @@ import { readFileSync } from 'fs'
 import globby from 'globby'
 import { join } from 'path'
 import { Annotation, buildEnvironment } from '../src'
-import { notEmpty } from '../src/extensions'
+import { List, notEmpty } from '../src/extensions'
 import validate from '../src/validator'
-import { Node, Problem } from './../src/model'
+import { Class, Literal, Node, Problem, Reference } from './../src/model'
 
 const TESTS_PATH = 'language/test/validations'
 
@@ -19,9 +19,9 @@ describe('Wollok Validations', () => {
   }))
   const environment = buildEnvironment(files)
 
-  const matchesExpectation = (problem: Problem, expected: Annotation) => {
+  const matchesExpectation = (problem: Problem, annotatedNode: Node, expected: Annotation) => {
     const code = expected.args['code']!
-    return problem.code === code && problem.sourceMap?.toString() === problem.node.sourceMap?.toString()
+    return problem.code === code && annotatedNode.sourceMap?.includes(problem.sourceMap!)
   }
 
   const errorLocation = (node: Node | Problem): string => `${node.sourceMap}`
@@ -48,26 +48,33 @@ describe('Wollok Validations', () => {
         const expectedProblems = allExpectations.get(node) || []
 
         for (const expectedProblem of expectedProblems) {
-          const code = expectedProblem.args['code']!
+          const code = expectedProblem.args['code']
           const level = expectedProblem.args['level']
+          const values = expectedProblem.args['values']
 
           if (!code) fail('Missing required "code" argument in @Expect annotation')
 
-          const errors = problems.filter(problem => !matchesExpectation(problem, expectedProblem))
+          const errors = problems.filter(problem => !matchesExpectation(problem, node, expectedProblem))
           if (notEmpty(errors))
             fail(`File contains errors: ${errors.map((_error) => _error.code + ' at ' + errorLocation(_error)).join(', ')}`)
 
-          const effectiveProblem = problems.find(problem => matchesExpectation(problem, expectedProblem))
+          const effectiveProblem = problems.find(problem => matchesExpectation(problem, node, expectedProblem))
           if (!effectiveProblem)
             fail(`Missing expected ${code} ${level ?? 'problem'} at ${errorLocation(node)}`)
 
 
           if (level && effectiveProblem.level !== level)
             fail(`Expected ${code} to be ${level} but was ${effectiveProblem.level} at ${errorLocation(node)}`)
+
+          if (values) {
+            const stringValues = (values as [Reference<Class>, List<Literal<string>>])[1].map(v => v.value)
+            if (stringValues.join('||') !== effectiveProblem.values.join('||'))
+              fail(`Expected ${code} to have ${JSON.stringify(stringValues)} but was ${JSON.stringify(effectiveProblem.values)} at ${errorLocation(node)}`)
+          }
         }
 
         for (const problem of problems) {
-          if (!expectedProblems.some(expectedProblem => matchesExpectation(problem, expectedProblem)))
+          if (!expectedProblems.some(expectedProblem => matchesExpectation(problem, node, expectedProblem)))
             fail(`Unexpected ${problem.code} ${problem.level} at ${errorLocation(node)}`)
         }
       })