+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## v4.0.5
+**Wollok Version: `3.1.9`**
+
+- Update Wollok version.
+- Fixed #159 and #164 - Singleton abstract methods validation
+- Fixed #172 - Filtering wollok base frames from stack trace & adding constants
+
+
+## v4.0.4
+**Wollok Version: `3.1.8`**
+
+- Update Wollok version.
+- Fixed validation `shouldNotDuplicateEntities` for programs and globals. https://github.com/uqbar-project/wollok-ts/pull/157
+- Fixed `malformedSentence` error for return methods. https://github.com/uqbar-project/wollok-ts/pull/162
+- Split validation `shouldInitializeAllAttributes` in two. https://github.com/uqbar-project/wollok-ts/pull/161
+
+
 ## v4.0.3
 **Wollok Version: `3.1.7`**